refactor(evalenv): address PR review feedback
