+++ envergo/moulinette/tests/test_evalenv.py
@@ -1,4 +1,5 @@
 import pytest
+from django.test import Client
 from django.urls import reverse
 from pytest_django.asserts import assertTemplateUsed
 
@@ -10,6 +11,7 @@ from envergo.moulinette.tests.factories import (
     RegulationFactory,
 )
 from envergo.moulinette.tests.utils import COORDS_BIZOU, make_amenagement_data
+from envergo.users.tests.factories import UserFactory
 
 
 def _bizou_data(created_surface, **extra):
@@ -379,3 +381,135 @@ def test_evalenv_subtractive_actions_to_take():
     assert actions_to_take_flatten == {
         "petitioner": ["depot_etude_impact"]
     }  # there is no pc_etude_impact in DB
+
+
+# ---------------------------------------------------------------------------
+# ICPE criterion
+# ---------------------------------------------------------------------------
+
+ICPE_BASE_PARAMS = "created_surface=500&final_surface=500&lng=-1.54394&lat=47.21381"
+
+
+@pytest.fixture
+def staff_client():
+    user = UserFactory(is_staff=True)
+    client = Client()
+    client.force_login(user)
+    return client
+
+
+@pytest.fixture
+def icpe_criterion(france_map):
+    ConfigAmenagementFactory(is_activated=True)
+    regulation = RegulationFactory(regulation="eval_env")
+    return CriterionFactory(
+        title="Cas par cas pour une installation classée (ICPE)",
+        regulation=regulation,
+        evaluator="envergo.moulinette.regulations.evalenv.ICPE",
+        activation_map=france_map,
+        is_optional=True,
+        is_staff_only=True,
+    )
+
+
+@pytest.mark.usefixtures("icpe_criterion")
+class TestICPEResults:
+    def _get_result(self, staff_client, icpe_projet, icpe_regime):
+        params = (
+            f"{ICPE_BASE_PARAMS}"
+            f"&evalenv_icpe-activate=on"
+            f"&evalenv_icpe-icpe_projet={icpe_projet}"
+            f"&evalenv_icpe-icpe_regime={icpe_regime}"
+        )
+        url = f"{reverse('moulinette_result')}?{params}"
+        return staff_client.get(url)
+
+    def test_creation_enregistrement_is_cas_par_cas(self, staff_client):
+        res = self._get_result(staff_client, "creation", "enregistrement")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "cas par cas" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_cas_par_cas.html")
+
+    def test_modif_avec_pac_enregistrement_is_cas_par_cas(self, staff_client):
+        res = self._get_result(staff_client, "modif_avec_pac", "enregistrement")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "cas par cas" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_cas_par_cas.html")
+
+    def test_creation_declaration_is_non_soumis(self, staff_client):
+        res = self._get_result(staff_client, "creation", "declaration")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "non soumis" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_non_soumis_declaration.html")
+
+    def test_modif_sans_pac_enregistrement_is_non_soumis(self, staff_client):
+        res = self._get_result(staff_client, "modif_sans_pac", "enregistrement")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "non soumis" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_non_soumis_sans_pac.html")
+
+    def test_aucun_aucun_is_non_soumis(self, staff_client):
+        res = self._get_result(staff_client, "aucun", "aucun")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "non soumis" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_non_soumis_pas_icpe.html")
+
+    def test_creation_inconnu_is_a_verifier(self, staff_client):
+        res = self._get_result(staff_client, "creation", "inconnu")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "à vérifier" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_a_verifier_creation.html")
+
+    def test_modif_avec_pac_inconnu_is_a_verifier(self, staff_client):
+        res = self._get_result(staff_client, "modif_avec_pac", "inconnu")
+        content = res.content.decode()
+
+        assert res.status_code == 200
+        assert "à vérifier" in content.lower()
+        assertTemplateUsed(res, "moulinette/eval_env/icpe_a_verifier_modification.html")
+
+
+@pytest.mark.usefixtures("icpe_criterion")
+class TestICPEFormValidation:
+    def _get_result(self, staff_client, icpe_projet, icpe_regime):
+        params = (
+            f"{ICPE_BASE_PARAMS}"
+            f"&evalenv_icpe-activate=on"
+            f"&evalenv_icpe-icpe_projet={icpe_projet}"
+            f"&evalenv_icpe-icpe_regime={icpe_regime}"
+        )
+        url = f"{reverse('moulinette_result')}?{params}"
+        return staff_client.get(url)
+
+    def test_aucun_projet_with_regime_is_invalid(self, staff_client):
+        """aucun projet + enregistrement regime = contradictory, redirects to form."""
+        res = self._get_result(staff_client, "aucun", "enregistrement")
+
+        assert res.status_code == 302
+        assert "/formulaire/" in res["Location"]
+
+    def test_projet_with_aucun_regime_is_invalid(self, staff_client):
+        """creation projet + aucun regime = contradictory, redirects to form."""
+        res = self._get_result(staff_client, "creation", "aucun")
+
+        assert res.status_code == 302
+        assert "/formulaire/" in res["Location"]
+
+    def test_coherent_answers_are_valid(self, staff_client):
+        """creation + enregistrement = valid, shows result page."""
+        res = self._get_result(staff_client, "creation", "enregistrement")
+
+        assert res.status_code == 200
+        assertTemplateUsed(res, "moulinette/result.html")
