+++ envergo/moulinette/tests/test_staff_only_criteria.py
@@ -91,115 +91,14 @@ class TestICPEStaffOnlyVisibility:
         assert res.status_code == 200
         assert "Installation classée (ICPE)" not in res.content.decode()
 
-    def test_superuser_non_staff_cannot_see_icpe_form(self, client):
+    def test_superuser_non_staff_can_see_icpe_form(self, client):
         user = UserFactory(is_superuser=True, is_staff=False)
         client.force_login(user)
         url = reverse("moulinette_form")
         res = client.get(url)
 
         assert res.status_code == 200
-        assert "Installation classée (ICPE)" not in res.content.decode()
-
-
-class TestICPEResults:
-    def _get_result(self, staff_client, icpe_projet, icpe_regime):
-        params = (
-            f"{BASE_PARAMS}"
-            f"&evalenv_icpe-activate=on"
-            f"&evalenv_icpe-icpe_projet={icpe_projet}"
-            f"&evalenv_icpe-icpe_regime={icpe_regime}"
-        )
-        url = f"{reverse('moulinette_result')}?{params}"
-        return staff_client.get(url)
-
-    def test_creation_enregistrement_is_cas_par_cas(self, staff_client):
-        res = self._get_result(staff_client, "creation", "enregistrement")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "cas par cas" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_cas_par_cas.html")
-
-    def test_modif_avec_pac_enregistrement_is_cas_par_cas(self, staff_client):
-        res = self._get_result(staff_client, "modif_avec_pac", "enregistrement")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "cas par cas" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_cas_par_cas.html")
-
-    def test_creation_declaration_is_non_soumis(self, staff_client):
-        res = self._get_result(staff_client, "creation", "declaration")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "non soumis" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_non_soumis_declaration.html")
-
-    def test_modif_sans_pac_enregistrement_is_non_soumis(self, staff_client):
-        res = self._get_result(staff_client, "modif_sans_pac", "enregistrement")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "non soumis" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_non_soumis_sans_pac.html")
-
-    def test_aucun_aucun_is_non_soumis(self, staff_client):
-        res = self._get_result(staff_client, "aucun", "aucun")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "non soumis" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_non_soumis_pas_icpe.html")
-
-    def test_creation_inconnu_is_a_verifier(self, staff_client):
-        res = self._get_result(staff_client, "creation", "inconnu")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "à vérifier" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_a_verifier_creation.html")
-
-    def test_modif_avec_pac_inconnu_is_a_verifier(self, staff_client):
-        res = self._get_result(staff_client, "modif_avec_pac", "inconnu")
-        content = res.content.decode()
-
-        assert res.status_code == 200
-        assert "à vérifier" in content.lower()
-        assertTemplateUsed(res, "moulinette/eval_env/icpe_a_verifier_modification.html")
-
-
-class TestICPEFormValidation:
-    def test_aucun_projet_with_regime_is_invalid(self, staff_client):
-        """aucun projet + enregistrement regime = contradictory, redirects to form."""
-        res = self._get_result(staff_client, "aucun", "enregistrement")
-
-        assert res.status_code == 302
-        assert "/formulaire/" in res["Location"]
-
-    def test_projet_with_aucun_regime_is_invalid(self, staff_client):
-        """creation projet + aucun regime = contradictory, redirects to form."""
-        res = self._get_result(staff_client, "creation", "aucun")
-
-        assert res.status_code == 302
-        assert "/formulaire/" in res["Location"]
-
-    def test_coherent_answers_are_valid(self, staff_client):
-        """creation + enregistrement = valid, shows result page."""
-        res = self._get_result(staff_client, "creation", "enregistrement")
-
-        assert res.status_code == 200
-        assertTemplateUsed(res, "moulinette/result.html")
-
-    def _get_result(self, staff_client, icpe_projet, icpe_regime):
-        params = (
-            f"{BASE_PARAMS}"
-            f"&evalenv_icpe-activate=on"
-            f"&evalenv_icpe-icpe_projet={icpe_projet}"
-            f"&evalenv_icpe-icpe_regime={icpe_regime}"
-        )
-        url = f"{reverse('moulinette_result')}?{params}"
-        return staff_client.get(url)
+        assert "Installation classée (ICPE)" in res.content.decode()
 
 
 class TestStaffOnlyFiltering:
